requirements.txt
streamlit==1.37.0
tensorflow==2.15.0
pillow
numpy
plotly
